Add ToString to Excel-related types for easier debugging
