+++ src/Lumina/Excel/Collection{T}.Enumerator.cs
