+++ src/Lumina/Excel/Collection{T}.cs
