+++ src/Lumina/Excel/Sheets/ExcelSheet{T}.cs
@@ -48,6 +48,9 @@ public readonly partial struct ExcelSheet< T > : IExcelSheet, ICollection< T >, 
     public Language Language => RawSheet.Language;
 
     /// <inheritdoc/>
+    public string Name => RawSheet.Name;
+
+    /// <inheritdoc/>
     public ExcelVariant Variant => RawSheet.Variant;
 
     /// <inheritdoc/>
@@ -155,6 +158,9 @@ public readonly partial struct ExcelSheet< T > : IExcelSheet, ICollection< T >, 
     IEnumerator< T > IEnumerable< T >.GetEnumerator() => GetEnumerator();
 
     IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
+
+    /// <inheritdoc/>
+    public override string ToString() => $"{Name}<{typeof( T ).Name}>({Language}, {Variant}, {Count} row(s), {Columns.Count} column(s))";
 
     /// <summary>Creates a row at the given index, without checking for bounds or preconditions.</summary>
     /// <param name="rowIndex">Index of the desired row.</param>
